requirements.txt
asn1crypto==0.24.0
cffi==1.12.2
Click==7.0
cryptography==2.6.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.8.1
pycparser==2.19
PyJWT==1.7.1
six==1.12.0
SQLAlchemy==1.3.2
Werkzeug==0.15.2
